+++ LoRaL2.h
@@ -15,11 +15,11 @@ public:
 	void operator=(const LoRaL2Packet&) = delete;
 	LoRaL2Packet() = delete;
 
-	LoRaL2Packet(uint8_t *packet, int len, int rssi, int err);
+	LoRaL2Packet(uint8_t *packet, size_t len, int rssi, int err);
 	virtual ~LoRaL2Packet();
 	
 	uint8_t *packet;
-	int len;
+	size_t len;
 	int rssi;
 	int err;
 };
@@ -32,25 +32,25 @@ public:
 	void operator=(const LoRaL2&) = delete;
 	
 	LoRaL2(long int band, int spread, int bandwidth,
-		const char *key, int key_len, recv_callback recv_cb);
+		const char *key, size_t key_len, recv_callback recv_cb);
 	virtual ~LoRaL2();
 	
-	bool send(const uint8_t *packet, int payload_len);
+	bool send(const uint8_t *packet, size_t payload_len);
 	uint32_t speed_bps() const;
-	int max_payload() const;
+	size_t max_payload() const;
 	bool ok() const;
 	// public because LoRa C API needs to call them
-	void on_recv(int rssi, uint8_t* packet, int len);
+	void on_recv(int rssi, uint8_t* packet, size_t len);
 	void on_sent();
 
 	/* private */
 	void resume_rx();
-	uint8_t *encrypt(const uint8_t *packet, int len, int& new_len);
-	uint8_t *append_fec(const uint8_t *packet, int len, int& new_len);
-	uint8_t *decode_fec(const uint8_t *packet, int len, int& new_len, int& err);
-	uint8_t *decrypt(const uint8_t *packet, int len, int& new_len, int& err);
+	uint8_t *encrypt(const uint8_t *packet, size_t len, size_t& new_len);
+	uint8_t *append_fec(const uint8_t *packet, size_t len, size_t& new_len);
+	uint8_t *decode_fec(const uint8_t *packet, size_t len, size_t& new_len, int& err);
+	uint8_t *decrypt(const uint8_t *packet, size_t len, size_t& new_len, int& err);
 	static uint8_t *hashed_key(const char* key, size_t len);
-	static void gen_iv(uint8_t* buffer, int len);
+	static void gen_iv(uint8_t* buffer, size_t len);
 
 	long int band;
 	int spread;
