+++ Radio.h
@@ -13,6 +13,6 @@ bool lora_start(long int band, int spread, int bandwidth, int power,
 		int paboost, int cr4slsh, LoRaL2 *pobserver);
 void lora_receive();
 bool lora_begin_packet();
-void lora_finish_packet(const uint8_t* packet, int len);
+void lora_finish_packet(const uint8_t* packet, size_t len);
 
 #endif
